cmds/pci: Move majority of cmds/pci into new pkg/pci
Move the majority of cmds/pci into new package pci to allow for future re-use of code by other commands.
