+++ pkg/pci/types.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package main
+package pci
 
 import "fmt"
 
@@ -20,6 +20,8 @@ type PCI struct {
 	DeviceName string
 }
 
+// String concatenates PCI address, Vendor, and Device to make a useful 
+// display for the user.
 func (p *PCI) String() string {
 	if *numbers {
 		return fmt.Sprintf("%s: %s:%s", p.Addr, p.Vendor, p.Device)
